requirements.txt
web3==6.2.0
requests==2.31.0
python-dotenv==1.0.0